requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
osmnx
tqdm
Pillow>=9.5.0
huggingface_hub>=0.24.0
datasets>=2.19.0